New packages for every subsystem and opmodes. Inside each package, classes for each subsystem. todo missions for driving mecanum wheels and lifting and lowering the arm
