+++ FtcRobotController-master/TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Subsystems/Arm.java
@@ -1,0 +1,8 @@
+package org.firstinspires.ftc.teamcode.Subsystems;
+
+public class Arm {
+
+    //TODO method that moves the arm to the highest point by the press of a button. by the press of the same button again, lowers it down.
+
+}
+
